requirements.txt
amqp==2.5.1
astroid==2.2.5
backports.csv==1.0.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.5
django-celery-beat==1.5.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-import-export==1.2.0
django-timezone-field==3.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
jdcal==1.4.1
kombu==4.6.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
monotonic==1.5
more-itertools==7.2.0
numpy==1.17.2
odfpy==1.4.0
openpyxl==2.6.3
pandas==0.25.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pylint==2.3.1
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.4.0
urllib3==1.25.4
vine==1.3.0
whitenoise==4.1.3
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
